// packages/json-joy/src/json-patch/codec/compact/types.ts
export type OPCODE_TEST_TYPE = OPCODE.test_type | 'test_type';
// ...
export type CompactTestTypeOp = [opcode: OPCODE_TEST_TYPE, path: string | Path, type: JsonPatchTypes[]];